Revert to ImageMagick 6
ImageMagick is a breaking API change from 6 to 7. As it seems no
Linux distros are really shipping 7 yet, it's unlikely that things
will compile easily against 7.

The easiest answer is to revert to ImageMagick 6 and keep on
rebuilding things that depend on it, each time it's updated.

This fixes tribblix/build#5